pyproject.toml
[project]
name = "ecommerce-pipeline"
version = "0.1.0"
description = "E-Commerce Polyglot Data Pipeline — Capstone Project"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    # Postgres
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    # MongoDB
    "pymongo>=4.6.0",
    # Redis (Phase 2)
    "redis>=5.0.0",
    # Neo4j (Phase 3)
    "neo4j>=5.18.0",
    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",   # FastAPI test client
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
seed = "scripts.seed:main"
migrate = "scripts.migrate:main"
setup = "scripts.setup:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ecommerce_pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
